test: Add a test for compdb parsing & move stuff.
