+++ README.md
@@ -2,3 +2,5 @@ WIP skeleton of a C++ indexer for faster iteration
 and understanding potential roadblocks.
 
 Might be potentially merged somewhere upstream in the future.
+
+See [CONTRIBUTING.md](./CONTRIBUTING.md) for build instructions etc.
