+++ indexer/BUILD
@@ -1,24 +1,24 @@
-cc_binary(
-    name = "scip-clang",
+cc_library(
+    name = "scip-clang-lib",
     srcs = [
         "CompilationDatabase.cc",
         "CompilationDatabase.h",
         "Driver.cc",
         "Driver.h",
+        "FileSystem.h",
         "IpcMessages.cc",
         "IpcMessages.h",
         "JsonIpcQueue.cc",
         "JsonIpcQueue.h",
         "LLVMAdapter.h",
-        "LLVMCommandLineParsing.h",
         "LLVMCommandLineParsing.cc",
+        "LLVMCommandLineParsing.h",
         "Logging.cc",
         "Logging.h",
-        "FileSystem.h",
         "Worker.cc",
         "Worker.h",
-        "main.cc",
     ],
+    visibility = ["//visibility:public"],
     deps = [
         "@com_google_absl//absl/algorithm:container",
         "@com_google_absl//absl/container:flat_hash_set",
@@ -29,28 +29,19 @@ cc_binary(
         "@spdlog",
         "@rapidjson",
         "@llvm-project//llvm:Support",
-        # FIXME: Figure out how to not build any backends at all.
-        # Right now, this requires all different backends to be
-        # compiled and linked. Looks like there is a variable
-        # LLVM_TARGETS_TO_BUILD which controls this
+        # FIXME: The tooling target seems to require at least
+        # one backend to be set. Figure out if we can somehow
+        # not build any backends at all, which would save on
+        # build time.
         "@llvm-project//clang:tooling",
-        # "@llvm-project//clang:ast",
     ],
 )
 
-# load("//tools:clang.bzl", "clang_tool")
-load("@com_grail_bazel_compdb//:defs.bzl", "compilation_database")
-load("@com_grail_bazel_output_base_util//:defs.bzl", "OUTPUT_BASE")
-
-# clang_tool("clang-format")
-# clang_tool("clang-tidy")
-# clang_tool("opt")
-
-compilation_database(
-    name = "compdb",
-    testonly = True,
-    output_base = OUTPUT_BASE,
-    targets = [
-        "//indexer:scip-clang",
+cc_binary(
+    name = "scip-clang",
+    srcs = [
+        "main.cc",
     ],
+    visibility = ["//tools:__pkg__"],
+    deps = [":scip-clang-lib"],
 )
