+++ test/compdb/simple-3.snapshot.yaml
@@ -1,0 +1,23 @@
+---
+- command:
+    - arg1
+    - arg2
+    - '-X=arg3'
+  file:            f1
+  directory:       d1
+  output:          ''
+- command:
+    - arg1
+    - arg2
+    - '-X="arg3"'
+  file:            f2
+  directory:       d2
+  output:          out2
+- command:
+    - '-Y=dq"dq'
+    - '-Z=blah'
+    - '-W=ww'
+  file:            f3
+  directory:       d3
+  output:          ''
+...
