+++ test/test_main.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+set -euo pipefail
+./test/test_main "$@"
