+++ tools/BUILD
@@ -1,0 +1,15 @@
+# TODO(def: cpp-linters): Add clang-format and clang-tidy usage here.
+
+load("@com_grail_bazel_compdb//:defs.bzl", "compilation_database")
+load("@com_grail_bazel_output_base_util//:defs.bzl", "OUTPUT_BASE")
+
+compilation_database(
+    name = "compdb",
+    testonly = True,
+    output_base = OUTPUT_BASE,
+    targets = [
+        "//indexer:scip-clang",
+        "//indexer:scip-clang-lib",
+        "//test:test_main",
+    ],
+)
