+++ tools/regen_compdb.sh
@@ -1,0 +1,10 @@
+#!/usr/bin/env bash
+bazel build //tools:compdb
+
+PROJECT_ROOT="$(dirname "${BASH_SOURCE[0]}")/.."
+
+if command -v jq &> /dev/null; then
+  jq . < "$PROJECT_ROOT/"bazel-bin/tools/compile_commands.json > compile_commands.json
+else
+  cp "$PROJECT_ROOT/"bazel-bin/tools/compile_commands.json compile_commands.json
+fi
